requirements.txt
pyserial==3.5
Pillow==9.5.0
cx_Freeze==6.15.9 